UltimosCambiosDefinitivo 07/04
No le den bola a los anteriores

+++ macowins/src/main/java/Grupo4/Sombrero.java
@@ -1,14 +1,14 @@
 package Grupo4;
 
 public class Sombrero implements Tipo {
-	private float coefMetro;
+	private double coefMetro;
 
 	
-	public Sombrero(float coefMetro) {
+	public Sombrero(double coefMetro) {
 		this.coefMetro = coefMetro;
 	}
 
-	public float precioBase(){
+	public double precioBase(){
 		return 150+(1+coefMetro);
 	}
 }
